requirements.txt
django>=5.1.7
openai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.4.2
